python: add resource list interface
Problem: The SchedResourceList interface only exists in the flux-resource
utility, so no other programs can share the code.

Move SchedResourceList from flux-resource.py to the bindings and add a
list_resources() function wrapper around it.

Call list_resources() in flux-resource.py and remove SchedResourceList
locally.
